# Conflicts:
#	app/src/main/java/com/umc/product/di/ApiModule.kt
#	app/src/main/java/com/umc/product/di/AuthenticationInterceptor.kt
#	app/src/main/java/com/umc/product/di/RepositoryModule.kt
#	app/src/main/java/com/umc/product/di/UseCaseModule.kt
#	data/src/main/java/com/umc/data/api/Endpoints.kt
#	data/src/main/java/com/umc/data/dataSource/local/AppDataStore.kt
#	data/src/main/java/com/umc/data/dataSource/remote/member/MemberRemoteDataSourceImpl.kt

+++ domain/src/main/java/com/umc/domain/usecase/schedule/GetScheduleDetailUseCase.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 class GetScheduleDetailUseCase @Inject constructor(
     private val scheduleRepository: ScheduleRepository
 ) {
-    suspend operator fun invoke(scheduleId: Int): ApiState<UserCheckAvailable> {
+    suspend operator fun invoke(scheduleId: Long): ApiState<UserCheckAvailable> {
         return scheduleRepository.getScheduleDetail(scheduleId)
     }
 }
